Add extended IRR support of some description.
This is a re-commit and force-push incorporating two previous commits.

Since "Unicode, Inc. specifically excludes the right to re-distribute
this file [GB12345.TXT] directly to third parties or other organizations
whether for profit or not." This is different from the usual terms, and
even from GB2312.TXT, where one is allowed to "make copies of this file
in any form for internal or external distribution as long as this notice
remains attached."

Since GB12345.TXT does not seem ever to have been updated to different
terms (and, indeed, the entire EASTASIA/GB/ directory is no longer being
hosted by Unicode), it seems I cannot include GB12345.TXT in this repo.