HKSCS support.
Commentary: I expected Big5 to be messy, and indeed it is proving to be.
As a sidenote, I've as a consequence of this basically finished up rolling
my own EUC encoding for HKSCS. Dunno if anyone else has been crazy enough
to do so. If so, it wouldn't necessarily be compatible.